Move xdg functions into its own package
So that I can re-use it in other parts of the code not related
to mutagen.
